Drew - Extracted UseAuth hook to a separate file, and general code cleanup
