+++ agrogo/src/components/AppContainer/AppContainer.tsx
@@ -1,7 +1,7 @@
 import "../../stylesheets/AppContainer.css";
 import Dashboard from "../Dashboard/Dashboard";
 import AuthenticationPopup from "../Auth/AuthenticationPopup";
-import { useAuth } from "../contexts/authContext/authentication";
+import { useAuth } from "../../hooks/UseAuth.tsx";
 
 function AppContainer() {
     const { userLoggedIn } = useAuth();
