+++ agrogo/src/components/contexts/authContext/authentication.tsx
@@ -1,37 +1,18 @@
-import React, { useContext, useEffect, useState } from "react";
+import { useEffect, useState, type ReactNode } from "react";
 import { auth } from "../../firebase/firebase"; // this allows our useEffect hook below to "listen" for when a user signs in or out
-import { onAuthStateChanged } from "firebase/auth";
+import { onAuthStateChanged, type User } from "firebase/auth";
+import { type valueUser, AuthContext } from "../../../hooks/UseAuth";
 
-// type of the value of props that will go into the context
-interface valueUser {
-    currentUser;
-    userLoggedIn: boolean;
-    loading: boolean;
-}
-
-const tempUser: valueUser = {currentUser: null, userLoggedIn: false, loading: true};
-
-// in React, "context" is a way to pass variables (like whether a user is logged in and who) while avoiding "props drilling"
-// props drilling is when you have to manually pass info as props from one component down a long line of other components, which gets difficult to keep track of very quickly
-// React lets you create your own context so that variables such as authentication can be easily used across comonents
-// we will initialize context here, and then use a Provider at the end of this function to set the values that will be shared as context
-const AuthContext = React.createContext(tempUser);
-
-// we're also exporting a custom hook, called useAuth
-// in React, hooks are the logic-only counterparts to components - you build them in a file somewhere, and then you can import that function/logic into other components to use and re-use later
-// some hooks, like useEffect, are built in, but you can also create your own hooks
-// even though we're exporting another hook here, and we're not really "writing" our own logic, this is still technically a custom hook that we'll use to get our new context moved around
-// so, we'll feed it the AuthContext we initialized above, and now we can export it and import it into other files
-// react hooks are weird - feel free to ask Madeline if you need clarification
-export function useAuth() {
-    return useContext(AuthContext);
+// Type of the value of props that will go into the Auth provider
+interface valueAuthProvider {
+    children: ReactNode;
 }
 
 // *** this is an important React concept - data is passed between components from parents to children as "props", handed off as arguments in the component function
 // we wrap our app in this component so that the whole app is a child, and can thus share this authentication state across all components
-export function AuthProvider({ children }) {
+export function AuthProvider({ children }: valueAuthProvider) {
     // these are how we handle the state of whether a user is logged in, who the current user is, and whether the app is still loading this info
-    const [currentUser, setCurrentUser] = useState(null);
+    const [currentUser, setCurrentUser] = useState<User | null>(null);
     const [userLoggedIn, setUserLoggedIn] = useState(false);
     const [loading, setLoading] = useState(true);
 
@@ -46,7 +27,7 @@ export function AuthProvider({ children }) {
     // async functions are used when your function depends on something that will take a longer-than-usual time to complete, usually because you're getting info from somewhere outside of your app
     // they're common with the useEffect hook for this reason, because the useEffect hook tells React that it needs to render everything first, then whatever is inside useEffect will happen
     // in our case, we're telling React that we want to render the component, then we can start expecting whatever info that we're getting from our Firebase API that is passing the auth object
-    async function initializeUser(user) {
+    async function initializeUser(user: User | null) {
         if(user) {
             // if we get a user, setting state with our current user
             // the "..." in js/ts is either the "rest parameter" or the "spread" operater; the spread operator "spreads" elements out so you can modify or combine them (more info here: https://www.geeksforgeeks.org/javascript/javascript-ellipsis/)
